requirements.txt
pytest==3.0.6
requests
flake8
PyYAML
python-jenkins==0.4.11
